changed number to i64 mapping

+++ workspace/model_generators/rust_serde_generator/src/lib.rs
@@ -529,7 +529,7 @@ impl TypeMapping {
     pub fn with_basic_types() -> Self {
         Self::empty()
             .add("string", "String")
-            .add("number", "f64")
+            .add("number", "i64")
             .add("integer", "i64")
             .add("boolean", "bool")
     }
